requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
pydantic==2.5.0
python-dotenv==1.0.0

# LLM
groq==0.4.1

# Web Search
tavily-python==0.3.0

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
readability-lxml==0.8.1
lxml==4.9.3

# Data Analysis & Visualization
matplotlib==3.8.2
plotly==5.18.0
pandas==2.1.3
numpy==1.26.2

# PDF Generation
markdown==3.5.1
weasyprint==60.1
reportlab==4.0.7

# Vector Store
chromadb==0.4.18

# Utilities
aiohttp==3.9.1
tenacity==8.2.3

